Remove another Bug in TCA
Is seems that sometimes a faulty positive PWM output is generated when a compare unit gets enabled "shortly" after the timer started. We can avoid that faulty ouput if we wait at least one complete period before enabling the first compare unit. I do not understand why this is needed, but it seems to fix the bug.